Keep remote port probe internal
